Factor out creature-editor.tsx.
